+++ tests/Notifier/TerminalNotifierNotifierTest.php
@@ -62,22 +62,6 @@ CLI;
     /**
      * {@inheritdoc}
      */
-    protected function getExpectedCommandLineForNotificationWithASubtitle()
-    {
-        return $this->getExpectedCommandLineForNotificationWithATitle();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getExpectedCommandLineForNotificationWithASound()
-    {
-        return $this->getExpectedCommandLineForNotificationWithATitle();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     protected function getExpectedCommandLineForNotificationWithAnIcon()
     {
         if (OsHelper::isMacOS() && version_compare(OsHelper::getMacOSVersion(), '10.9.0', '>=')) {