+++ tests/Notifier/ToasterNotifierTest.php
@@ -62,22 +62,6 @@ CLI;
     /**
      * {@inheritdoc}
      */
-    protected function getExpectedCommandLineForNotificationWithASubtitle()
-    {
-        return $this->getExpectedCommandLineForNotificationWithATitle();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getExpectedCommandLineForNotificationWithASound()
-    {
-        return $this->getExpectedCommandLineForNotificationWithATitle();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     protected function getExpectedCommandLineForNotificationWithAnIcon()
     {
         $iconDir = $this->getIconDir();